Support more than 64k concurrent clients
The total number of TCP connections to a single destinatiation IP, port tuple
is limited to the hosts ephemeral TCP port range (on Linux
/proc/sys/net/ipv4/ip_local_port_range). To enable more concurrent connections
to a single destinatiation we can bind to multiple source addresses. This patch
permits this by introducing a -s argument which can be used to specify source
addresses for clients to use. Clients are distributed between the specified
source addresses. In addition this patch changes the type of the client count
variables from uint16_t to uint32_t to support more 65535 concurrent clients.

Signed-off-by: Dustin Lundquist <dustin@null-ptr.net>