ignore aborted requests when benchmark ends early
ignore aborted requests when benchmark ends early,
instead of treating those requests as having failed

However, any request in progress that has taken more than cutoff seconds
(default 4 seconds) is counted as having failed so that hung connections
are counted as failed.  If "normal" responses take more than cutoff
seconds, those aborted requests may be counted as failed, too.

cutoff can be defined with -s <secs>, and the default is 4 seconds
(similarities to Apache Bench (ab) '-s')

Note: Ctrl-C or weighttp -a <secs> can be used to cause the benchmark
to end early, with the full time accounted (for any connections waiting
for a response for more than cutoff seconds).  Hung connections are not
terminated at cutoff, but rather are accounted for at end of benchmark.