do not adjust req_started if tests end early
The number of started requests may be more than the number of completed
responses if the tests end early.  The number of status codes reported,
e.g. 2xx, may exceed the number of completed responses if response
headers were received but response body still in progress when the tests
ended.

The timing statistics attempt to omit the timings for responses
abandoned at the end of the tests, and are based on number of retired
requests reported.